Handler works as check.
If no check is implemented, consider return value of handler as result of the
check. The idea is that e.g. <done /> doesn't require separate check, because
the success of action should be in check (but the action already
happened/failed to happen). This behaviour may be reverted in future.