Fix getsibling behaviour.
Previously, when node itself was in filtered nodes of syblings, the counting
didn't work correctly. Now, it should work same way as it wouldn't be in
the filtered list (siblings).
vector == 0 => The node itself if it matches criteria, None otherwise
vector != 0 => sibling in the specified distance before or after the node