Add support for fail_reason
When fail_reason attribute of an element is specified, its value
is checked against fail_reason value of a result when a check result
is 'fail'. If a fail is expected and the values are equal, the
overall result is 'pass', otherwise the overall result is 'fail'.
When fail_reason is not specified, the failure reason of a result
is not taken into consideration.