Add support for dogtail node liveness checks
A new exception class 'NonexistentError' was added as a parent of
TimeoutError, introducing a more general exception that should be
raised in a case where a dogtail node in question either was not
found (it doesn't exist anymore or has never existed) or is dead.

The existing functions dealing with dogtail nodes were modified
to check whether a node in question is alive and showing (visible),
which should prevent anabot from getting stuck when a method
is called on a dead dogtail node (e. g. a dialog that doesn't
exist anymore). NonexistentError exception is raised when the
queried node is dead and not showing.