Wait for network to be up before fetching Anabot recipe
It's not guaranteed that network is up when 05-download_recipe-preexec.hook is
executed and if it's not up, it fails fetching the recipe. Since it's
practically impossible to currently use Anabot without network, just add a new
hook which just waits from NetworkManager to signal that the network is fully
operational.

Introduction of this hook could possibly also mean, that following commit
could be reverted: 2e432a55087e7d2a5ca628778a10bdea5289aa2c