Add family of asserts for alternate valid texts
Sometimes there might be more than one correct texts, so these asserts
allow passing lists of (translated) texts.

This correspond to Python's `in` test, hence the naming convention.

In order to keep the library consistent, we're adding `In` assert for
every `Equals` assert, except where it does not make sense:

 *  `assertPasswordTextInputEquals()` is already more complex than
    equality check, and the assert is about a text created by the
    test; it's exceedingly unlikely that we would want to accept
    known alternatives.

 *  `assertCheckboxEquals()` is about a discrete (often binary) state
    so similar point applies as above: it's unlikely that test knows
    about multiple equivalent states.