[RTT-2432] Don't mix in mandatory groups to anabot-packageset.txt
The purpose of anabot-packageset.txt is to record user's selection of
environment and addons.

Listing environment *and* its mandatory groups is redundant, so let's
only list environment.  This is, after all, what anaconda does when it's
constructing kickstart.  (I've checked this on RHEL-8 and RHEL-9.)

This reverts cc7a246a from 5 years ago, where mandatory group list was
added.  However, it's not completely clear why the behavior was changed
(Not even after discussion with the original author---note that this was
before RTT was using Jira for tracking the development so there is no
additional thread).  I'm going to assume this was an attempt to to match
old anaconda / yum behavior; in that case it's not relevant now.