[RTT-4715] Ensure that language is entered properly
It seems that the whole Anaconda application is not focused at first
under some circumstances. This may cause that the language is not
entered properly in the search field and subsequently make Anabot crash
when it tries to click on a language cell that is out of the visible area.

This fix ensures that the language is found and therefore the
language cell is displayed in the visible area.