[RTT-4707] Workaround for proper lang ids for Chinese
Anabot uses langtable to get the best guess for language id
based on the language/locale selection on the Welcome screen.
However, for Chinese, langtable returns 'zh_XXXX_YY', which
technically might be correct, but the available translations
for Anaconda are 'zh_YY'.

No matter if this is a bug in the translation identifiers or not,
we need to have this working, so that we can test even installations
in Chinese and this workaround can be reverted in the future
if there's a proper fix implemented.