Update base handler for advanced partitioning details
So far the base handler called default_handler() with a fixed
local node (page tab containing details about selected device).
The local node may, however, become invalid under some circumstances,
for example when a LUKS device is unlocked.

To fix this, it's needed to check that the local node is still
valid before executing a particular handler/check, and update
it to a new one if needed.