[RTT-5714] Add temporary hack to get a virtual monitor in RDP mode
The hack consists of running a local RDP connection via xfreerdp,
which results in creating a virtual monitor with a defined resolution.
Anabot is not able to work properly without it in RDP mode, as
dogtail tracebacks due to some reportedly negative widget coordinates
or dimensions.

It is supposed to be only a temporary workaround and should be removed
when there's a proper solution (when RHEL-64694 is fixed).