Ensure there's always an opened GDK display present
Sometimes it happens that there's no GDK display present and this
leads to a traceback when a keyCombo() is sent. It's not entirely
sure if this is a bug in dogtail or if we should open the display
explicitly ourselves. We can remove this piece of code in the future
if we find out that this should rather get fixed in dogtail.