[RTT-6267] Add workaround for Reboot button traceback
Sometimes a DBusException('The name is not activatable') exception
occurs when Anabot clicks on the Reboot button. It's not sure why
at this point, but at least it can be worked around in case a proper
fix won't be found, as it happens at the very end of the
installation process.