Translate 'push button' to 'button' if necessary
A new version of at-spi2 renamed the button ATK node type from
'push button' to 'button', thus we need to take this into account
and translate the calls to make the existing code work.

The ideal solution would be to use node types from Atspi.Role enum
instead of the string identifiers, but refactoring the code doesn't
look to be worth the effort.